move pages into app
