+++ app/[locale]/(app)/program/page.tsx
@@ -1,4 +1,4 @@
-import { ProgramPage } from "@/components/program/ProgramPage";
+import { ProgramPage } from "./ProgramPage";
 import { client } from "@/sanity/lib/client";
 import { PROGRAMPAGE_QUERY } from "@/sanity/lib/queries/programPage";
 import { PROGRAMPAGE_QUERYResult } from "@/sanity/types/types";
@@ -13,7 +13,5 @@ export default async function Page({
     lang,
   });
 
-  if (data) return <ProgramPage data={data} lang={lang} />;
-
-  return null;
+  return <ProgramPage data={data} />;
 }
