+++ app/[locale]/(event)/program/[slug]/page.tsx
@@ -1,4 +1,4 @@
-import { EventPage } from "@/components/event/EventPage";
+import { EventPage } from "@/app/[locale]/(event)/program/[slug]/EventPage";
 import { sanityFetch } from "@/sanity/lib/live";
 import { EVENT_QUERY } from "@/sanity/lib/queries/event";
 import { EVENT_QUERYResult } from "@/sanity/types/types";
@@ -15,10 +15,6 @@ export default async function Page({
     query: EVENT_QUERY,
     params: { lang, slug },
   });
-
-  if (!data) {
-    return null;
-  }
 
   return <EventPage data={data} />;
 }
